requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alpaca-trade-api==2.3.0
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///tmp/build/80754af9/argon2-cffi-bindings_1644569681205/work
async-timeout==4.0.2
asynctest==0.13.0
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///opt/conda/conda-bld/beautifulsoup4_1650462163268/work
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
Bottleneck @ file:///tmp/build/80754af9/bottleneck_1648028896657/work
brotlipy==0.7.0
certifi==2022.6.15
cffi @ file:///tmp/build/80754af9/cffi_1636541929935/work
charset-normalizer==2.0.12
click==8.1.2
cloudpickle==2.0.0
cryptography @ file:///tmp/build/80754af9/cryptography_1652083456434/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
daal4py==2021.5.0
dask==2022.2.0
debugpy @ file:///tmp/build/80754af9/debugpy_1637091426235/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
deprecation==2.1.0
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649908585034/work
fastjsonschema @ file:///tmp/build/80754af9/python-fastjsonschema_1620414857593/work/dist
fastparquet @ file:///tmp/build/80754af9/fastparquet_1632742478800/work
fonttools==4.25.0
frozenlist==1.3.0
fsspec==2022.3.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1648562407465/work
importlib-resources @ file:///tmp/build/80754af9/importlib_resources_1625135880749/work
inflection==0.5.1
ipykernel @ file:///tmp/build/80754af9/ipykernel_1646983213839/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1643818144432/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1634143127070/work
jedi @ file:///tmp/build/80754af9/jedi_1644299024593/work
Jinja2 @ file:///opt/conda/conda-bld/jinja2_1647436528585/work
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
jsonschema @ file:///tmp/build/80754af9/jsonschema_1650008095244/work
jupyter==1.0.0
jupyter-client @ file:///opt/conda/conda-bld/jupyter_client_1650622202839/work
jupyter-console @ file:///opt/conda/conda-bld/jupyter_console_1647002188872/work
jupyter-core @ file:///tmp/build/80754af9/jupyter_core_1646994442130/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
kiwisolver @ file:///opt/conda/conda-bld/kiwisolver_1638569886207/work
llvmlite==0.38.0
locket==1.0.0
lxml==4.9.1
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1621528142364/work
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1647441664166/work
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mistune @ file:///tmp/build/80754af9/mistune_1594373098390/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
more-itertools @ file:///tmp/build/80754af9/more-itertools_1637733554872/work
msgpack==1.0.3
multidict==6.0.2
multitasking==0.0.11
munkres==1.1.4
nbclient @ file:///tmp/build/80754af9/nbclient_1650308370114/work
nbconvert @ file:///opt/conda/conda-bld/nbconvert_1649751911790/work
nbformat @ file:///tmp/build/80754af9/nbformat_1649844918397/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1649847911453/work
notebook @ file:///tmp/build/80754af9/notebook_1644985983866/work
numba @ file:///opt/conda/conda-bld/numba_1648040517072/work
numerapi==2.11.0
numexpr @ file:///tmp/build/80754af9/numexpr_1640704205123/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1649782762302/work
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.3.4
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parquet==1.3.1
parquet-cli==1.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
partd==1.2.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.0.1
ply==3.11
prometheus-client @ file:///opt/conda/conda-bld/prometheus_client_1643788673601/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyarrow==8.0.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1636098896055/work
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-snappy @ file:///home/conda/feedstock_root/build_artifacts/python-snappy_1616178114734/work
pytz==2022.1
PyYAML==6.0
pyzmq @ file:///tmp/build/80754af9/pyzmq_1638434924971/work
qtconsole @ file:///opt/conda/conda-bld/qtconsole_1649078897110/work
QtPy @ file:///opt/conda/conda-bld/qtpy_1649073884068/work
Quandl @ file:///tmp/build/80754af9/quandl_1614799137919/work
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1642601761909/work
scikit-learn-intelex==2021.20220215.162512
scipy @ file:///tmp/build/80754af9/scipy_1641536880743/work
seaborn @ file:///tmp/build/80754af9/seaborn_1629307859561/work
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
soupsieve @ file:///tmp/build/80754af9/soupsieve_1636706018808/work
ta==0.10.1
terminado @ file:///tmp/build/80754af9/terminado_1644304459949/work
testpath @ file:///tmp/build/80754af9/testpath_1624638946665/work
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
thrift==0.11.0
thriftpy2==0.4.14
toolz==0.11.2
tornado @ file:///tmp/build/80754af9/tornado_1606942283357/work
tqdm==4.64.0
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing_extensions==4.2.0
urllib3 @ file:///tmp/abs_5dhwnz6atv/croots/recipe/urllib3_1659110457909/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client==1.3.3
websockets==10.3
widgetsnbextension @ file:///tmp/build/80754af9/widgetsnbextension_1645009354203/work
yarl==1.7.2
yfinance==0.1.74
zipp==3.8.0
